refactor(christmas): 테스트 클래스 분리, 필드명 변경
